+++ adapters/AppLovin/AppLovinAdapter/GADMAdapterAppLovinConstant.m
@@ -12,8 +12,6 @@ NSString *const GADMAdapterAppLovinErrorDomain = @"com.google.mediation.applovin
 
 NSString *const GADMAdapterAppLovinSDKErrorDomain = @"com.google.mediation.applovinSDK";
 
-NSString *const GADMAdapterAppLovinInfoPListSDKKey = @"AppLovinSdkKey";
-
 NSString *const GADMAdapterAppLovinAdapterVersion = @"11.11.3.0";
 
 NSString *const GADMAdapterAppLovinSDKKey = @"sdkKey";