+++ adapters/AppLovin/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## AppLovin iOS Mediation Adapter Changelog
 
+#### Next Version
+- Adapter no longer checks for the AppLovin SDK key from the `Info.plist` file.
+
 #### [Version 11.11.3.0](https://dl.google.com/googleadmobadssdk/mediation/ios/applovin/AppLovinAdapter-11.11.3.0.zip)
 - Verified compatibility with AppLovin SDK 11.11.3.